backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]>=2.0.36
aiosqlite==0.20.0
asyncpg==0.29.0
pydantic-settings==2.2.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.9
celery==5.4.0
redis==5.0.8
stripe==9.9.0
python-pptx==0.6.23
httpx==0.27.0
psycopg2-binary==2.9.9
alembic==1.13.2
anthropic>=0.34.0
openai>=1.40.0
Pillow>=10.3.0
# Testing
pytest==8.3.2
pytest-asyncio==0.23.8
anyio==4.4.0
